+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/ChangeDoneFragment.kt
@@ -2,11 +2,10 @@ package au.com.shiftyjelly.pocketcasts.account
 
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
 import androidx.core.os.bundleOf
 import androidx.fragment.app.activityViewModels
+import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.account.viewmodel.DoneViewModel
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
@@ -29,18 +28,18 @@ class ChangeDoneFragment : BaseFragment() {
     private val shouldCloseParent: Boolean
         get() = arguments?.getBoolean(ARG_CLOSE_PARENT) ?: false
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                AppThemeWithBackground(theme.activeTheme) {
-                    ChangeDonePage(
-                        viewModel = viewModel,
-                        closeForm = {
-                            closeForm()
-                        },
-                    )
-                }
-            }
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ) = content {
+        AppThemeWithBackground(theme.activeTheme) {
+            ChangeDonePage(
+                viewModel = viewModel,
+                closeForm = {
+                    closeForm()
+                },
+            )
         }
     }