+++ modules/features/reimagine/src/main/kotlin/au/com/shiftyjelly/pocketcasts/reimagine/podcast/SharePodcastFragment.kt
@@ -10,10 +10,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.ComposeView
 import androidx.core.os.BundleCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.viewModels
+import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.reimagine.ui.ShareColors
@@ -57,22 +57,20 @@ class SharePodcastFragment : BaseDialogFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ) = ComposeView(requireActivity()).apply {
-        val platforms = SocialPlatform.getAvailablePlatforms(requireContext())
-        setContent {
-            val assetController = rememberBackgroundAssetControler(shareColors)
-            val listener = remember { shareListenerFactory.create(this@SharePodcastFragment, assetController, args.source) }
+    ) = content {
+        val platforms = remember { SocialPlatform.getAvailablePlatforms(requireContext()) }
+        val assetController = rememberBackgroundAssetControler(shareColors)
+        val listener = remember { shareListenerFactory.create(this@SharePodcastFragment, assetController, args.source) }
 
-            val uiState by viewModel.uiState.collectAsState()
-            SharePodcastPage(
-                podcast = uiState.podcast,
-                episodeCount = uiState.episodeCount,
-                socialPlatforms = platforms,
-                shareColors = shareColors,
-                assetController = assetController,
-                listener = listener,
-            )
-        }
+        val uiState by viewModel.uiState.collectAsState()
+        SharePodcastPage(
+            podcast = uiState.podcast,
+            episodeCount = uiState.episodeCount,
+            socialPlatforms = platforms,
+            shareColors = shareColors,
+            assetController = assetController,
+            listener = listener,
+        )
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {