+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/EpisodeArtworkConfigurationFragment.kt
@@ -2,7 +2,6 @@ package au.com.shiftyjelly.pocketcasts.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -15,11 +14,12 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
+import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRow
@@ -36,30 +36,29 @@ import au.com.shiftyjelly.pocketcasts.ui.R as UR
 
 @AndroidEntryPoint
 class EpisodeArtworkConfigurationFragment : BaseFragment() {
-    @Inject lateinit var settings: Settings
+    @Inject
+    lateinit var settings: Settings
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View = ComposeView(requireContext()).apply {
-        val sortedElements = ArtworkConfiguration.Element.entries.sortedBy { getString(it.titleId) }
-        setContent {
-            AppThemeWithBackground(theme.activeTheme) {
-                val artworkConfiguration by settings.artworkConfiguration.flow.collectAsState()
+    ) = content {
+        val sortedElements = remember { ArtworkConfiguration.Element.entries.sortedBy { getString(it.titleId) } }
+        AppThemeWithBackground(theme.activeTheme) {
+            val artworkConfiguration by settings.artworkConfiguration.flow.collectAsState()
 
-                EpisodeArtworkSettings(
-                    artworkConfiguration = artworkConfiguration,
-                    elements = sortedElements,
-                    onUpdateConfiguration = { configuration ->
-                        settings.artworkConfiguration.set(configuration, updateModifiedAt = true)
-                    },
-                    onBackPressed = {
-                        @Suppress("DEPRECATION")
-                        activity?.onBackPressed()
-                    },
-                )
-            }
+            EpisodeArtworkSettings(
+                artworkConfiguration = artworkConfiguration,
+                elements = sortedElements,
+                onUpdateConfiguration = { configuration ->
+                    settings.artworkConfiguration.set(configuration, updateModifiedAt = true)
+                },
+                onBackPressed = {
+                    @Suppress("DEPRECATION")
+                    activity?.onBackPressed()
+                },
+            )
         }
     }
 
@@ -128,14 +127,15 @@ class EpisodeArtworkConfigurationFragment : BaseFragment() {
         )
     }
 
-    private val ArtworkConfiguration.Element.titleId get() = when (this) {
-        ArtworkConfiguration.Element.Filters -> LR.string.filters
-        ArtworkConfiguration.Element.UpNext -> LR.string.up_next
-        ArtworkConfiguration.Element.Downloads -> LR.string.profile_navigation_downloads
-        ArtworkConfiguration.Element.Files -> LR.string.profile_navigation_files
-        ArtworkConfiguration.Element.Starred -> LR.string.profile_navigation_starred
-        ArtworkConfiguration.Element.Bookmarks -> LR.string.bookmarks
-        ArtworkConfiguration.Element.ListeningHistory -> LR.string.profile_navigation_listening_history
-        ArtworkConfiguration.Element.Podcasts -> LR.string.podcasts
-    }
+    private val ArtworkConfiguration.Element.titleId
+        get() = when (this) {
+            ArtworkConfiguration.Element.Filters -> LR.string.filters
+            ArtworkConfiguration.Element.UpNext -> LR.string.up_next
+            ArtworkConfiguration.Element.Downloads -> LR.string.profile_navigation_downloads
+            ArtworkConfiguration.Element.Files -> LR.string.profile_navigation_files
+            ArtworkConfiguration.Element.Starred -> LR.string.profile_navigation_starred
+            ArtworkConfiguration.Element.Bookmarks -> LR.string.bookmarks
+            ArtworkConfiguration.Element.ListeningHistory -> LR.string.profile_navigation_listening_history
+            ArtworkConfiguration.Element.Podcasts -> LR.string.podcasts
+        }
 }