+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/SettingsFragment.kt
@@ -2,7 +2,6 @@ package au.com.shiftyjelly.pocketcasts.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
@@ -10,9 +9,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rxjava2.subscribeAsState
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
+import androidx.fragment.compose.content
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -42,45 +41,42 @@ class SettingsFragment : BaseFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View =
-        ComposeView(requireContext()).apply {
-            setContent {
-                AppThemeWithBackground(theme.activeTheme) {
-                    var isUnrestrictedBattery by remember { mutableStateOf(batteryRestrictions.isUnrestricted()) }
-                    DisposableEffect(this) {
-                        val observer = LifecycleEventObserver { _, event ->
-                            if (event == Lifecycle.Event.ON_RESUME) {
-                                isUnrestrictedBattery = batteryRestrictions.isUnrestricted()
-                            }
-                        }
-
-                        lifecycle.addObserver(observer)
-                        onDispose {
-                            lifecycle.removeObserver(observer)
-                        }
+    ) = content {
+        AppThemeWithBackground(theme.activeTheme) {
+            var isUnrestrictedBattery by remember { mutableStateOf(batteryRestrictions.isUnrestricted()) }
+            DisposableEffect(this) {
+                val observer = LifecycleEventObserver { _, event ->
+                    if (event == Lifecycle.Event.ON_RESUME) {
+                        isUnrestrictedBattery = batteryRestrictions.isUnrestricted()
                     }
+                }
 
-                    userManager
-                        .getSignInState()
-                        .subscribeAsState(null)
-                        .value
-                        ?.let { signInState ->
-                            val bottomInset = settings.bottomInset.collectAsStateWithLifecycle(0)
-                            SettingsFragmentPage(
-                                signInState = signInState,
-                                onBackPressed = {
-                                    @Suppress("DEPRECATION")
-                                    activity?.onBackPressed()
-                                },
-                                isDebug = BuildConfig.DEBUG,
-                                isUnrestrictedBattery = isUnrestrictedBattery,
-                                openFragment = { fragment ->
-                                    (activity as? FragmentHostListener)?.addFragment(fragment)
-                                },
-                                bottomInset = bottomInset.value.pxToDp(LocalContext.current).dp,
-                            )
-                        }
+                lifecycle.addObserver(observer)
+                onDispose {
+                    lifecycle.removeObserver(observer)
                 }
             }
+
+            userManager
+                .getSignInState()
+                .subscribeAsState(null)
+                .value
+                ?.let { signInState ->
+                    val bottomInset = settings.bottomInset.collectAsStateWithLifecycle(0)
+                    SettingsFragmentPage(
+                        signInState = signInState,
+                        onBackPressed = {
+                            @Suppress("DEPRECATION")
+                            activity?.onBackPressed()
+                        },
+                        isDebug = BuildConfig.DEBUG,
+                        isUnrestrictedBattery = isUnrestrictedBattery,
+                        openFragment = { fragment ->
+                            (activity as? FragmentHostListener)?.addFragment(fragment)
+                        },
+                        bottomInset = bottomInset.value.pxToDp(LocalContext.current).dp,
+                    )
+                }
         }
+    }
 }