+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 7.78
------
+-----*   New Features
+    *   Add episode basic authentication support
+        ([#3213](https://github.com/Automattic/pocket-casts-android/pull/3213))
+*   Bug Fixes
+    *   Fix effects being lost when paused
+        ([#3219](https://github.com/Automattic/pocket-casts-android/pull/3219))
 *   Updates
+    *   Update profile header UI
+        ([#3235](https://github.com/Automattic/pocket-casts-android/pull/3235))
     *   Deferred addition to UpNext post update after sync 
         ([#3211](https://github.com/Automattic/pocket-casts-android/pull/312))